+++ spec/lib/karafka/web/tracking/consumers/contracts/consumer_group_spec.rb
@@ -1,0 +1,82 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  subject(:contract) { described_class.new }
+
+  let(:consumer_group) do
+    {
+      id: 'example_app_karafka_web',
+      subscription_groups: {
+        'c81e728d9d4c_1' => {
+          id: 'c81e728d9d4c_1',
+          state: {
+            state: 'up',
+            join_state: 'steady',
+            stateage: 90_002,
+            rebalance_age: 90_000,
+            rebalance_cnt: 1,
+            rebalance_reason: 'Metadata for subscribed topic(s) has changed'
+          },
+          topics: {
+            'karafka_consumers_reports' => {
+              name: 'karafka_consumers_reports',
+              partitions: {
+                0 => {
+                  lag_stored: 0,
+                  lag_stored_d: 0,
+                  committed_offset: 18,
+                  stored_offset: 18,
+                  fetch_state: 'active',
+                  id: 0,
+                  poll_state: 'active'
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  end
+
+  context 'when config is valid' do
+    it { expect(contract.call(consumer_group)).to be_success }
+  end
+
+  context 'when id is missing' do
+    before { consumer_group.delete(:id) }
+
+    it { expect(contract.call(consumer_group)).not_to be_success }
+  end
+
+  context 'when id is empty' do
+    before { consumer_group[:id] = '' }
+
+    it { expect(contract.call(consumer_group)).not_to be_success }
+  end
+
+  context 'when id is not a string' do
+    before { consumer_group[:id] = 123 }
+
+    it { expect(contract.call(consumer_group)).not_to be_success }
+  end
+
+  context 'when subscription_groups is missing' do
+    before { consumer_group.delete(:subscription_groups) }
+
+    it { expect(contract.call(consumer_group)).not_to be_success }
+  end
+
+  context 'when subscription_groups is not a hash' do
+    before { consumer_group[:subscription_groups] = 'not a hash' }
+
+    it { expect(contract.call(consumer_group)).not_to be_success }
+  end
+
+  context 'when subscription_group does not have an id' do
+    let(:expected_error) { Karafka::Web::Errors::Tracking::ContractError }
+
+    before { consumer_group[:subscription_groups]['c81e728d9d4c_1'].delete(:id) }
+
+    it { expect { contract.call(consumer_group) }.to raise_error(expected_error) }
+  end
+end